Restructure create dao utils
