+++ apps/web/src/modules/create-dao/utils/calculateMaxAllocation.ts
@@ -1,0 +1,17 @@
+import { Duration } from 'src/typings'
+import { toSeconds } from 'src/utils/helpers'
+
+export const calculateMaxAllocation = (
+  allocation: string | number,
+  end: string | number,
+  auctionDuration: Duration
+) => {
+  const auctionDurationInSeconds = toSeconds(auctionDuration)
+  const endDate = new Date(end).getTime()
+  const now = new Date().getTime()
+  const diffInSeconds = Math.abs((endDate - now) / 1000)
+  const frequency = Number(allocation)
+  const numberOfAuctionsTilEndDate = diffInSeconds / auctionDurationInSeconds
+
+  return Math.floor(numberOfAuctionsTilEndDate * (frequency / 100))
+}
