+++ apps/web/src/modules/create-dao/utils/formatFounderAllocation.ts
@@ -1,0 +1,9 @@
+import { TokenAllocation } from 'src/typings'
+
+export const formatFounderAllocation = ({
+  founderAddress,
+  allocationPercentage,
+  endDate,
+}: TokenAllocation): string => {
+  return `${founderAddress} will receive ${allocationPercentage}% of Tokens, until ${endDate}.`
+}
