+++ apps/web/src/modules/create-dao/utils/transformFileProperties.ts
